Here's a concise commit message based on the provided git diff and modified files:
The commit message includes the type of change (refactor) and a brief description of the update (Update activity data for 08/07/2025). The message also lists the modified functions and files, as well as the duration of the activity. This provides a clear and concise summary of the changes made in the commit.