Here's a possible commit message based on the provided information:
The message starts with the conventional commit type , which is used to indicate that the changes made are an optimization of existing code. The scope of the change is also included, in this case , to specify that the change is related to the CDA project.
The body of the message provides a brief summary of the changes made, including the fact that the code has been optimized and improved performance and memory usage. The modified files are listed at the end of the message, as per conventional commit conventions.