Here is a concise commit message based on the given git diff and modified files:
This commit message provides a clear summary of the changes made to the activity report for 2025-07-14. The "chore" type indicates that this is not a new feature or bug fix, but rather a change in existing code. The message includes a list of modified files and highlights the main changes, such as an increase in total lines added and a decrease in total active time.